troquei de long pra int o tipo de id para não dar erro no deletar mentor

+++ src/main/java/org/scam/entities/CoordenacaoEntity.java
@@ -10,7 +10,7 @@ public class CoordenacaoEntity implements UsuarioEntity{
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY) // SERIAL no PostgreSQL = IDENTITY
-    private Long id;
+    private int id;
 
     @Column(name = "nome", nullable = false, length = 45)
     private String nome;
@@ -26,7 +26,7 @@ public class CoordenacaoEntity implements UsuarioEntity{
     }
 
     //construtor com todos os campos
-    public CoordenacaoEntity(Long id, String nome, String email, String senha){
+    public CoordenacaoEntity(int id, String nome, String email, String senha){
         this.id = id;
         this.nome = nome;
         this.email = email;
@@ -34,8 +34,8 @@ public class CoordenacaoEntity implements UsuarioEntity{
     }
 
     //getters e setters
-    public Long getId(){return id;}
-    public void setId(Long id){this.id = id;}
+    public int getId(){return id;}
+    public void setId(int id){this.id = id;}
 
     public String getNome(){return nome;}
     public void setNome(String nome){this.nome = nome;}
